Create index, more functions & glue everything 🚀
The main highlight is the
`doTheMagicMongooseSchemaIntoJSONObjectsAndTSTypeDefinitionFilesParsing`
function which does exactly what it's name says :D

Obviously, I didn't just create this massive file out of nowhere. This
fella took a lot of hours to complete (I started the `tongoose` project
just 5 days ago and here we are!, but it took me exactly 37 hours
(tracked by wakatime) to complete - which is quite an insane amount of
time for me - but I absolutely loved making this project! I really hope
that my dear tongoose will bring value to someone c:)

Signed-off-by: Kipras Melnikovas <sarpixo@gmail.com>